requirements.txt
asgiref==3.7.2
astroid==3.0.1
asttokens==2.4.0
backcall==0.2.0
boto3==1.28.57
botocore==1.31.57
crispy-bootstrap4==2022.1
decorator==5.1.1
dill==0.3.7
Django==4.2.5
django-crispy-forms==2.0
django-multiselectfield==0.1.12
django-storages==1.14.1
django-tempus-dominus==5.1.2.17
djangorestframework==3.14.0
exceptiongroup==1.1.3
executing==1.2.0
ipdb==0.13.13
ipython==8.16.0
isort==5.12.0
jedi==0.19.0
jmespath==1.0.1
matplotlib-inline==0.1.6
mccabe==0.7.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pylint==3.0.2
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
tomlkit==0.12.2
traitlets==5.10.1
typing_extensions==4.8.0
urllib3==1.26.16
wcwidth==0.2.7
